Implementação dos testes da camada de serviço de Renda varíavel.

+++ src/main/java/io/github/pameladilly/service/impl/RendaVariavelImpl.java
@@ -57,7 +57,12 @@ public class RendaVariavelImpl implements RendaVariavelService {
     @Override
     public void excluir(RendaVariavel rendaVariavel) {
 
+        if(! repository.findById(rendaVariavel.getIdAtivo()).isPresent()) {
+            throw new RendaVariavelNotFound();
+        }
+
         repository.delete(rendaVariavel);
+
     }
 
     @Override
